Eleva selectTab para index
